Replace form component with config
